Move default summary code into Cell
